#252 fix: code review + style
