+++ src/Template/Layout/js/app/helpers/download.js
@@ -1,0 +1,47 @@
+import Vue from 'vue';
+
+export default {
+    install (Vue, options) {
+        Vue.prototype.$helpers = {
+            /**
+            * Force download using a syntetic element
+            *
+            * @param {*} blob
+            * @param {*} filename
+            */
+            forceDownload(blob, filename) {
+                let a = document.createElement('a');
+                a.download = filename;
+                a.href = blob;
+                a.click();
+            },
+
+            /**
+            * download a resource as a blob to avoid cors restrictions
+            *
+            * @param {string} url
+            * @param {string} filename
+            */
+            downloadResource(url, filename) {
+                if (!filename) {
+                    filename = url.split('\\').pop().split('/').pop();
+                }
+
+                const options = {
+                    headers: new Headers({
+                        'Origin': location.origin
+                    }),
+                    mode: 'cors'
+                }
+
+                fetch(url, options)
+                    .then(response => response.blob())
+                    .then(blob => {
+                        let blobUrl = window.URL.createObjectURL(blob);
+                        this.forceDownload(blobUrl, filename);
+                    })
+                    .catch(e => console.error(e));
+            }
+        }
+    }
+};
