+++ src/Template/Layout/js/app/pages/modules/view.js
@@ -10,6 +10,7 @@ import PropertyView from 'app/components/property-view/property-view';
 import RelationView from 'app/components/relation-view/relation-view';
 
 export default {
+
     components: {
         PropertyView,
     },
@@ -39,45 +40,5 @@ export default {
         toggleTabs() {
             return this.tabsOpen = !this.tabsOpen;
         },
-
-        /**
-         * Force download using a syntetic element
-         *
-         * @param {*} blob
-         * @param {*} filename
-         */
-        forceDownload(blob, filename) {
-            let a = document.createElement('a');
-            a.download = filename;
-            a.href = blob;
-            a.click();
-        },
-
-        /**
-         * download a resource as a blob to avoid cors restrictions
-         *
-         * @param {string} url
-         * @param {string} filename
-         */
-        downloadResource(url, filename) {
-            if (!filename) {
-                filename = url.split('\\').pop().split('/').pop();
-            }
-
-            const options = {
-                headers: new Headers({
-                    'Origin': location.origin
-                }),
-                mode: 'cors'
-            }
-
-            fetch(url, options)
-                .then(response => response.blob())
-                .then(blob => {
-                    let blobUrl = window.URL.createObjectURL(blob);
-                    this.forceDownload(blobUrl, filename);
-                })
-                .catch(e => console.error(e));
-        }
     }
 }
